Store instance root only in the InstancePresenter now
